MPS-4981 [build:3180] Exception when editing

+++ core/languageDesign/editor/source/jetbrains/mps/nodeEditor/EditorComponent.java
@@ -904,7 +904,9 @@ public abstract class EditorComponent extends JComponent implements Scrollable, 
     if (myNodeSubstituteChooser != null) {
       myNodeSubstituteChooser.dispose();
     }
-    ((EditorCell_Basic) myRootCell).onRemove();
+    if (myRootCell != null) {
+      ((EditorCell_Basic) myRootCell).onRemove();
+    }
   }
 
   private void addOurListeners(SModelDescriptor sm) {
